fix: embedding async
- Updated Embedding Manager with async from query_manager
